Integrate ReplayEngine and add COMMON mode for transaction simulator
Major changes:
1. Integrated ReplayEngine into SimulatorApplication
   - Added COMMON mode for simple sequential processing
   - Added STREAMING mode support in ReplayEngine
   - Refactored existing executeSimulation() methods to use ReplayEngine

2. Enhanced ReplayEngine with new modes:
   - COMMON: Simple sequential processing without rate limiting
   - FIXED_QPS: Rate-limited processing with Guava RateLimiter
   - REPLAY_DT: Time-based replay with acceleration
   - SCENARIO: Scenario testing with max concurrency
   - STREAMING: High-concurrency reactive streaming

3. Unified transactionDt field type to Long
   - Changed from Double to Long across all modules
   - Updated CSV parsing, mapping, and replay logic
   - Added parseLong() method to CsvTransactionReader

4. Improved rate limiting in FIXED_QPS mode
   - Added tryAcquire with timeout to prevent blocking
   - Better handling of max in-flight requests

Configuration:
- Added COMMON to Mode enum in SimulatorProperties
- Updated application.yml with new mode options

+++ tools/txn-simulator/src/main/java/com/risk/sim/mapping/FeatureEncoder.java
@@ -32,7 +32,7 @@ public class FeatureEncoder {
 
         // TransactionDT (normalized to hours)
         if (record.getTransactionDt() != null) {
-            double dtHours = record.getTransactionDt() / 3600.0;
+            double dtHours = (double)(record.getTransactionDt()) / 3600.0;
             features.put("TransactionDT_hours", dtHours);
         }
