+++ tools/txn-simulator/src/main/java/com/risk/sim/source/TransactionRecord.java
@@ -30,7 +30,7 @@ public class TransactionRecord {
     private boolean isFraud;
 
     // Transaction fields
-    private Double transactionDt;
+    private Long transactionDt;
     private BigDecimal transactionAmt;
     private String productCd;  // (e.g., "W", "H", "C", "S", "R")
